backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Redis
redis==5.0.1
hiredis==2.2.3

# CORS
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
